requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
astroid==4.0.4
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
coverage==7.13.4
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
decord==0.6.0
dill==0.4.1
einops==0.8.1
exceptiongroup==1.3.1
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.1.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
iopath==0.1.10
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.6.1
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.12.0.88
opencv-python==4.13.0.92
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
ptflops==0.7.5
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
pyparsing==3.3.1
pysodmetrics==1.6.0
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
ruff==0.15.1
safetensors==0.7.0
sam3 @ file:///home/tyler/Projects/SAM3-DUNet/sam3-main
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.22
tokenizers==0.22.1
tomli==2.4.0
tomlkit==0.14.0
torch==2.10.0
torchaudio==2.10.0
torchmetrics==1.8.2
# Editable install with no version control (torchsig==2.1.0)
-e /mnt/c/Projects/torchsig-main
torchvision==0.25.0
tqdm==4.67.1
transformers==4.57.3
triton==3.6.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wandb==0.23.1
wcwidth==0.2.14
yarl==1.22.0
